del
Added del function which deletes keys from the model.
Will delete keys from objects or indices from arrays.
Refactored code to use a do_walk function that takes
an action which is invoked at the last key in the path.
The set, push and del functions then become very simple
and you just have to code the last action to be taken.